Finish

+++ ValidateUI/ValidateUI/PictureEx.cpp
@@ -525,7 +525,7 @@ BOOL CPictureEx::Load(LPCTSTR szFileName)
 		return FALSE;
 	};
 
-	dwSize = file.GetLength();
+	dwSize = (DWORD)file.GetLength();
 	hGlobal = GlobalAlloc(GMEM_MOVEABLE | GMEM_NODISCARD,dwSize);
 	if (!hGlobal)
 	{
